requirements.txt
alembic==1.5.7
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
MongoAlchemy==0.19
mongoengine==0.23.0
mysqlclient==2.0.3
PyJWT==1.4.2
pymongo==3.11.3
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
WTForms==2.3.3
